+++ routes/storeInfo/storeInfo.js
@@ -1,8 +1,6 @@
 var express = require('express');
 var router = express.Router();
 
-const hangul = require('hangul-js');
-
 const statusCode = require('../../module/utils/statusCode');
 const responseMessage = require('../../module/utils/responseMessage');
 const authUtil = require('../../module/utils/authUtil');
@@ -56,39 +54,6 @@ router.post('/', upload.single('thumbnail'), async(req, res) => {
 router.get('/', async(req, res) => {
     StoreInfo.readAll()
     .then(({ code, json }) => {
-        res.status(code).send(json);
-    })
-    .catch((err) => {
-        console.log(err);
-        res.status(statusCode.INTERNAL_SERVER_ERROR, authUtil.successFalse(statusCode.INTERNAL_SERVER_ERROR, responseMessage.INTERNAL_SERVER_ERROR));
-    });
-});
-
-
-/**
- * [GET] /storeInfo/search?keyword={타이핑 시 검색어}
- * 식당 이름 검색
- * @author 100yeeun
- * @params 검색키워드
- */
-
-router.get('/search', async(req, res) => {
-    StoreInfo.readAllName()
-    .then(({ result, code }) => {
-
-        const searcher = new hangul.Searcher(req.query.keyword);
-
-        const findStoreNameList = new Array();
-
-		for (var i =0 ; i<result.length; i++){
-			if (searcher.search(result[i].name)>=0){
-				findStoreNameList.push(result[i]); 
-			}
-		}
-
-        const json = authUtil.successTrue(statusCode.OK, responseMessage.X_READ_SUCCESS('검색'), findStoreNameList)
-
-        
         res.status(code).send(json);
     })
     .catch((err) => {
